Started improvements to pipelines, filters, parsers

+++ src/components/framework/pipeline/connectors/fileconnector/fileconnector.cpp
@@ -56,7 +56,7 @@ namespace qor{ namespace components{
 
     bool FileConnector::Connect()
     {
-        m_File = new_ref<platform::IFile>(*m_pFileIndex, m_OpenFor, m_Mode);
+        m_File = new_ref<platform::IFile>(*m_pFileIndex, static_cast<int>(m_OpenFor), static_cast<int>(m_Mode));
         m_connected = !m_File.IsNull();//->IsOpen();
         return m_connected;
     }
